RS bug fixes and some modification to the main dashboard

+++ app/Http/Controllers/HomeController.php
@@ -42,8 +42,12 @@ class HomeController extends Controller
         $a_devices_count = $devices->get()->where('active', 1)->count();
         $in_devices_count = $devices->get()->where('active', 0)->count();
 
-        $employee_info = $employee->all();
+        $employee_info = $employee->with('devices')->with('roles')->with('groups')->get();
         $device_info = $devices->all();
+        $d = $devices->with('employees')->orderBy('active', 'DESC')->get();
+        // foreach($d as $dd){
+        // dd($employee_info);
+        // }
 
 
         return view('home', [
@@ -52,7 +56,9 @@ class HomeController extends Controller
             'a_d_count' => $a_devices_count,
             'in_d_count' => $in_devices_count,
             'employee_info' => $employee_info,
-            'devices' => $device_info
+            'employee_device' => $d
+
+
         ]);
     }
 
@@ -113,19 +119,23 @@ class HomeController extends Controller
     }
     public function AssignEmployee(Request $request, employees $employees, employees_role $employees_role, employees_group $employees_group, employees_applications $employees_applications, employees_devices $employees_devices, devices $devices)
     {
+        $error = 1;
         if ($request->ajax()) {
 
             $employee_assignment_data = $employees->find($request->id);
 
             //ROLES
-            //Before inserting delete the old values
-            $delete_old_data = $employees_role->where('employees_id', $request->id)->forceDelete();
-            foreach ($request->roles as $role) {
-                $employees_role->create([
-                    'employees_id' => $request->id,
-                    'role_id' => $role
-                ]);
+            if ($request->roles != NULL) {
+                //Before inserting delete the old values
+                $delete_old_data = $employees_role->where('employees_id', $request->id)->forceDelete();
+                foreach ($request->roles as $role) {
+                    $employees_role->create([
+                        'employees_id' => $request->id,
+                        'role_id' => $role
+                    ]);
+                }
             }
+
             //GROUPS
             $delete_old_group_data = $employees_group->where('employees_id', $request->id)->forceDelete();
             $employees_group->employees_id = $request->id;
@@ -146,48 +156,84 @@ class HomeController extends Controller
                     ]);
                 }
             }
+
             //Devices
+            //IF THE DEVICE IS NONE REQUEST->DEVICE is 0 -> Then delete a value from employees_devices (if it exsits)
+            if ($request->device == 0) {
+                //BEFORE REMOVING LETS SEE IF THAT VALUE EVEN EXSITS
+                $count_employees_devices_table_count = $employees_devices->where('employees_id', $request->id)->count();
 
+                //IF THE RETURN VALUE IS ZERO THAT MEANS NO FURTHER ACTION IS NEEDED THE DATA WAS NEVER THERE
+                if ($count_employees_devices_table_count != 0) {
 
-            //get old device id
-            $old_device = $employees_devices->where('employees_id', $request->id)->first();
-            if ($old_device != NULL) {
-                if ($old_device->devices_id != $request->device) {
-                    //deactivate the old device
+                    $old_device = $employees_devices->where('employees_id', $request->id)->first();
+                    //WE ALSO WANT TO DEACTIVATE THE OLD ASSIGNED DEVICE
                     $devices->where('id', $old_device->devices_id)->update(['active' => 0]);
-                    //Before entering the new device check if it iis already assigned
+                    //THEN DELETE THE RECORD FROM EMPLOYEES_DEVICES TABLE
+                    $employees_devices->where('devices_id', $old_device->devices_id)->forceDelete();
+                    $error = 0;
+                } else {
+                    $error = 0;
+                }
+            } else {
+                $old_device = $employees_devices->where('employees_id', $request->id)->first();
+                if ($old_device != NULL) {
+                    if ($old_device->devices_id != $request->device) { // IF THERE IS A NEW DEVICE
+                        //NOW FOR ALL REQUESTS THAT ARE NONE ZERO
+                        //1st LETS CHECK IF THAT DEVICE IS ASSIGNED OR NOT
+                        $d_count = $employees_devices->where('devices_id', $request->device)->count();
+
+                        $errors = "Cannot assign this device because it has been already assigned";
+                        if ($d_count > 0) {
+                            return response()->json(['error' => $errors], 422);
+                        } else {
+                            $employees_count = $employees_devices->where('employees_id', $request->id)->count();
+                            if ($employees_count == 0) {
+                                //IF NOT LETS STORE THE INFOMATION IN TABLE EMPLOYEES_DEVICES AND ALSO ACTIVATE THE DEVICE IN DEVICES TABLE
+                                $employees_devices->employees_id = $request->id;
+                                $employees_devices->devices_id = $request->device;
+                                $employees_devices->save();
+                                //Update the last entry
+                                $devices->where('id', $request->device)->update(['active' => 1]);
+                                $error = 0;
+                            } else {
+                                $old_device = $employees_devices->where('employees_id', $request->id)->first();
+                                //WE ALSO WANT TO DEACTIVATE THE OLD ASSIGNED DEVICE
+                                $devices->where('id', $old_device->devices_id)->update(['active' => 0]);
+                                //WE NEED TO JUST UPDATE THE DEVICE VALUE AND DEACTIVATE THE OLD DEVICE
+                                $employees_devices->where('employees_id', $request->id)->update(['devices_id' => $request->device]);
+                                $devices->where('id', $request->device)->update(['active' => 1]);
+                                $error = 0;
+                            }
+                        }
+                    } else {
+                        $error = 0;
+                    }
+                } else { //there are no records found and the request is not zero
+                    //NOW FOR ALL REQUESTS THAT ARE NONE ZERO
+                    //1st LETS CHECK IF THAT DEVICE IS ASSIGNED OR NOT
                     $d_count = $employees_devices->where('devices_id', $request->device)->count();
 
                     $errors = "Cannot assign this device because it has been already assigned";
                     if ($d_count > 0) {
                         return response()->json(['error' => $errors], 422);
+                    } else {
+                        //IF NOT LETS STORE THE INFOMATION IN TABLE EMPLOYEES_DEVICES AND ALSO ACTIVATE THE DEVICE IN DEVICES TABLE
+                        $employees_devices->employees_id = $request->id;
+                        $employees_devices->devices_id = $request->device;
+                        $employees_devices->save();
+                        //Update the last entry
+                        $devices->where('id', $request->device)->update(['active' => 1]);
+                        $error = 0;
                     }
-                    $employees_devices->where('employees_id', $request->id)->forceDelete();
-                    $employees_devices->employees_id = $request->id;
-                    $employees_devices->devices_id = $request->device;
-                    $employees_devices->save();
-                    //Update the last entry
-                    $devices->where('id', $request->device)->update(['active' => 1]);
                 }
-            } else {
-                //Before entering the new device check if it iis already assigned
-                $d_count = $employees_devices->where('devices_id', $request->device)->count();
-
-                $errors = "Cannot assign this device because it has been already assigned";
-                if ($d_count > 0) {
-                    return response()->json(['error' => $errors], 422);
-                }
-                $employees_devices->where('employees_id', $request->id)->forceDelete();
-                $employees_devices->employees_id = $request->id;
-                $employees_devices->devices_id = $request->device;
-                $employees_devices->save();
-                //Update the last entry
-                $devices->where('id', $request->device)->update(['active' => 1]);
             }
-
-
-
-            $success_message = "User has been updated";
+            if ($error == 0) {
+                $success_message = "User has been updated";
+            } else {
+                $errors = "issues";
+                return response()->json(['error' => $errors], 422);
+            }
 
             return response()->json(['success' => $success_message]);
         }
